Started Sidebar, fixed accessToken error
created a Home and Sidebar component that the user will be taken to once access token is initiated. Every time the app was rendered again, it would take user to login screen again. Fixed this by checking for access token every time rendered so that it wouldn't start the access token initiated procedure again.