User top-items
Retrieved and stored top tracks and artists that the user listens to. The top three are listed in the home page. The profile page holds the list of the top 20 tracks listened to. Need to make a toggle that allows users to look at short, medium, and long term top items.