Radar Chart (chart.js)
Improved styling of the Listen.js component. Installed a working and accurate radar chart from chart.js. There is a conflict between the necessity of clearing the chart and when the song changes (therefore the stats change). Currently the chart spasms every 5 seconds. Will work on this soon. Other than that, its looking great.